p4raw-id: //depot/cfgperl@4887

+++ win32/win32.c
@@ -671,7 +671,7 @@ win32_opendir(char *filename)
     WIN32_FIND_DATAW	wFindData;
     HANDLE		fh;
     char		buffer[MAX_PATH*2];
-    WCHAR		wbuffer[MAX_PATH];
+    WCHAR		wbuffer[MAX_PATH+1];
     char*		ptr;
 
     len = strlen(filename);
@@ -942,9 +942,9 @@ remove_dead_process(long child)
     if (child >= 0) {
 	dTHXo;
 	CloseHandle(w32_child_handles[child]);
-	Copy(&w32_child_handles[child+1], &w32_child_handles[child],
+	Move(&w32_child_handles[child+1], &w32_child_handles[child],
 	     (w32_num_children-child-1), HANDLE);
-	Copy(&w32_child_pids[child+1], &w32_child_pids[child],
+	Move(&w32_child_pids[child+1], &w32_child_pids[child],
 	     (w32_num_children-child-1), DWORD);
 	w32_num_children--;
     }
@@ -969,9 +969,9 @@ remove_dead_pseudo_process(long child)
     if (child >= 0) {
 	dTHXo;
 	CloseHandle(w32_pseudo_child_handles[child]);
-	Copy(&w32_pseudo_child_handles[child+1], &w32_pseudo_child_handles[child],
+	Move(&w32_pseudo_child_handles[child+1], &w32_pseudo_child_handles[child],
 	     (w32_num_pseudo_children-child-1), HANDLE);
-	Copy(&w32_pseudo_child_pids[child+1], &w32_pseudo_child_pids[child],
+	Move(&w32_pseudo_child_pids[child+1], &w32_pseudo_child_pids[child],
 	     (w32_num_pseudo_children-child-1), DWORD);
 	w32_num_pseudo_children--;
     }
@@ -1036,7 +1036,7 @@ win32_stat(const char *path, struct stat *buffer)
     char	t[MAX_PATH+1]; 
     int		l = strlen(path);
     int		res;
-    WCHAR	wbuffer[MAX_PATH];
+    WCHAR	wbuffer[MAX_PATH+1];
     HANDLE      handle;
     int         nlink = 1;
 
@@ -1228,7 +1228,7 @@ DllExport char *
 win32_getenv(const char *name)
 {
     dTHXo;
-    WCHAR wBuffer[MAX_PATH];
+    WCHAR wBuffer[MAX_PATH+1];
     DWORD needlen;
     SV *curitem = Nullsv;
 
@@ -1392,11 +1392,13 @@ win32_unlink(const char *filename)
     DWORD attrs;
 
     if (USING_WIDE()) {
-	WCHAR wBuffer[MAX_PATH];
+	WCHAR wBuffer[MAX_PATH+1];
 
 	A2WHELPER(filename, wBuffer, sizeof(wBuffer));
 	wcscpy(wBuffer, PerlDir_mapW(wBuffer));
 	attrs = GetFileAttributesW(wBuffer);
+	if (attrs == 0xFFFFFFFF)
+	    goto fail;
 	if (attrs & FILE_ATTRIBUTE_READONLY) {
 	    (void)SetFileAttributesW(wBuffer, attrs & ~FILE_ATTRIBUTE_READONLY);
 	    ret = _wunlink(wBuffer);
@@ -1409,6 +1411,8 @@ win32_unlink(const char *filename)
     else {
 	filename = PerlDir_mapA(filename);
 	attrs = GetFileAttributesA(filename);
+	if (attrs == 0xFFFFFFFF)
+	    goto fail;
 	if (attrs & FILE_ATTRIBUTE_READONLY) {
 	    (void)SetFileAttributesA(filename, attrs & ~FILE_ATTRIBUTE_READONLY);
 	    ret = unlink(filename);
@@ -1419,6 +1423,9 @@ win32_unlink(const char *filename)
 	    ret = unlink(filename);
     }
     return ret;
+fail:
+    errno = ENOENT;
+    return -1;
 }
 
 DllExport int
@@ -1430,7 +1437,7 @@ win32_utime(const char *filename, struct utimbuf *times)
     FILETIME ftAccess;
     FILETIME ftWrite;
     struct utimbuf TimeBuffer;
-    WCHAR wbuffer[MAX_PATH];
+    WCHAR wbuffer[MAX_PATH+1];
 
     int rc;
     if (USING_WIDE()) {
@@ -2048,7 +2055,7 @@ DllExport FILE *
 win32_fopen(const char *filename, const char *mode)
 {
     dTHXo;
-    WCHAR wMode[MODE_SIZE], wBuffer[MAX_PATH];
+    WCHAR wMode[MODE_SIZE], wBuffer[MAX_PATH+1];
     FILE *f;
     
     if (!*filename)
@@ -2097,7 +2104,7 @@ DllExport FILE *
 win32_freopen(const char *path, const char *mode, FILE *stream)
 {
     dTHXo;
-    WCHAR wMode[MODE_SIZE], wBuffer[MAX_PATH];
+    WCHAR wMode[MODE_SIZE], wBuffer[MAX_PATH+1];
     if (stricmp(path, "/dev/null")==0)
 	path = "NUL";
 
@@ -2403,8 +2410,8 @@ win32_link(const char *oldname, const char *newname)
 {
     dTHXo;
     BOOL (__stdcall *pfnCreateHardLinkW)(LPCWSTR,LPCWSTR,LPSECURITY_ATTRIBUTES);
-    WCHAR wOldName[MAX_PATH];
-    WCHAR wNewName[MAX_PATH];
+    WCHAR wOldName[MAX_PATH+1];
+    WCHAR wNewName[MAX_PATH+1];
 
     if (IsWin95())
 	Perl_die(aTHX_ PL_no_func, "link");
@@ -2429,26 +2436,31 @@ win32_link(const char *oldname, const char *newname)
 DllExport int
 win32_rename(const char *oname, const char *newname)
 {
-    WCHAR wOldName[MAX_PATH];
-    WCHAR wNewName[MAX_PATH];
-    char szOldName[MAX_PATH];
+    WCHAR wOldName[MAX_PATH+1];
+    WCHAR wNewName[MAX_PATH+1];
+    char szOldName[MAX_PATH+1];
+    char szNewName[MAX_PATH+1];
     BOOL bResult;
+    dTHXo;
+
     /* XXX despite what the documentation says about MoveFileEx(),
      * it doesn't work under Windows95!
      */
     if (IsWinNT()) {
-	dTHXo;
+	DWORD dwFlags = MOVEFILE_COPY_ALLOWED;
 	if (USING_WIDE()) {
 	    A2WHELPER(oname, wOldName, sizeof(wOldName));
 	    A2WHELPER(newname, wNewName, sizeof(wNewName));
+	    if (wcsicmp(wNewName, wOldName))
+		dwFlags |= MOVEFILE_REPLACE_EXISTING;
 	    wcscpy(wOldName, PerlDir_mapW(wOldName));
-	    bResult = MoveFileExW(wOldName,PerlDir_mapW(wNewName),
-			MOVEFILE_COPY_ALLOWED|MOVEFILE_REPLACE_EXISTING);
+	    bResult = MoveFileExW(wOldName,PerlDir_mapW(wNewName), dwFlags);
 	}
 	else {
-	    strcpy(szOldName, PerlDir_mapA(szOldName));
-	    bResult = MoveFileExA(szOldName,PerlDir_mapA(newname),
-			MOVEFILE_COPY_ALLOWED|MOVEFILE_REPLACE_EXISTING);
+	    if (stricmp(newname, oname))
+		dwFlags |= MOVEFILE_REPLACE_EXISTING;
+	    strcpy(szOldName, PerlDir_mapA(oname));
+	    bResult = MoveFileExA(szOldName,PerlDir_mapA(newname), dwFlags);
 	}
 	if (!bResult) {
 	    DWORD err = GetLastError();
@@ -2473,14 +2485,17 @@ win32_rename(const char *oname, const char *newname)
     }
     else {
 	int retval = 0;
-	char tmpname[MAX_PATH+1];
+	char szTmpName[MAX_PATH+1];
 	char dname[MAX_PATH+1];
 	char *endname = Nullch;
 	STRLEN tmplen = 0;
 	DWORD from_attr, to_attr;
 
+	strcpy(szOldName, PerlDir_mapA(oname));
+	strcpy(szNewName, PerlDir_mapA(newname));
+
 	/* if oname doesn't exist, do nothing */
-	from_attr = GetFileAttributes(oname);
+	from_attr = GetFileAttributes(szOldName);
 	if (from_attr == 0xFFFFFFFF) {
 	    errno = ENOENT;
 	    return -1;
@@ -2490,7 +2505,7 @@ win32_rename(const char *oname, const char *newname)
 	 * don't delete it in case oname happens to be the same file
 	 * (but perhaps accessed via a different path)
 	 */
-	to_attr = GetFileAttributes(newname);
+	to_attr = GetFileAttributes(szNewName);
 	if (to_attr != 0xFFFFFFFF) {
 	    /* if newname is a directory, we fail
 	     * XXX could overcome this with yet more convoluted logic */
@@ -2498,29 +2513,29 @@ win32_rename(const char *oname, const char *newname)
 		errno = EACCES;
 		return -1;
 	    }
-	    tmplen = strlen(newname);
-	    strcpy(tmpname,newname);
-	    endname = tmpname+tmplen;
-	    for (; endname > tmpname ; --endname) {
+	    tmplen = strlen(szNewName);
+	    strcpy(szTmpName,szNewName);
+	    endname = szTmpName+tmplen;
+	    for (; endname > szTmpName ; --endname) {
 		if (*endname == '/' || *endname == '\\') {
 		    *endname = '\0';
 		    break;
 		}
 	    }
-	    if (endname > tmpname)
-		endname = strcpy(dname,tmpname);
+	    if (endname > szTmpName)
+		endname = strcpy(dname,szTmpName);
 	    else
 		endname = ".";
 
 	    /* get a temporary filename in same directory
 	     * XXX is this really the best we can do? */
-	    if (!GetTempFileName((LPCTSTR)endname, "plr", 0, tmpname)) {
+	    if (!GetTempFileName((LPCTSTR)endname, "plr", 0, szTmpName)) {
 		errno = ENOENT;
 		return -1;
 	    }
-	    DeleteFile(tmpname);
+	    DeleteFile(szTmpName);
 
-	    retval = rename(newname, tmpname);
+	    retval = rename(szNewName, szTmpName);
 	    if (retval != 0) {
 		errno = EACCES;
 		return retval;
@@ -2528,16 +2543,16 @@ win32_rename(const char *oname, const char *newname)
 	}
 
 	/* rename oname to newname */
-	retval = rename(oname, newname);
+	retval = rename(szOldName, szNewName);
 
 	/* if we created a temporary file before ... */
 	if (endname != Nullch) {
 	    /* ...and rename succeeded, delete temporary file/directory */
 	    if (retval == 0)
-		DeleteFile(tmpname);
+		DeleteFile(szTmpName);
 	    /* else restore it to what it was */
 	    else
-		(void)rename(tmpname, newname);
+		(void)rename(szTmpName, szNewName);
 	}
 	return retval;
     }
@@ -2567,7 +2582,7 @@ win32_open(const char *path, int flag, ...)
     dTHXo;
     va_list ap;
     int pmode;
-    WCHAR wBuffer[MAX_PATH];
+    WCHAR wBuffer[MAX_PATH+1];
 
     va_start(ap, flag);
     pmode = va_arg(ap, int);
@@ -2854,7 +2869,7 @@ win32_mkdir(const char *dir, int mode)
 {
     dTHXo;
     if (USING_WIDE()) {
-	WCHAR wBuffer[MAX_PATH];
+	WCHAR wBuffer[MAX_PATH+1];
 	A2WHELPER(dir, wBuffer, sizeof(wBuffer));
 	return _wmkdir(PerlDir_mapW(wBuffer));
     }
@@ -2866,7 +2881,7 @@ win32_rmdir(const char *dir)
 {
     dTHXo;
     if (USING_WIDE()) {
-	WCHAR wBuffer[MAX_PATH];
+	WCHAR wBuffer[MAX_PATH+1];
 	A2WHELPER(dir, wBuffer, sizeof(wBuffer));
 	return _wrmdir(PerlDir_mapW(wBuffer));
     }
@@ -2878,7 +2893,7 @@ win32_chdir(const char *dir)
 {
     dTHXo;
     if (USING_WIDE()) {
-	WCHAR wBuffer[MAX_PATH];
+	WCHAR wBuffer[MAX_PATH+1];
 	A2WHELPER(dir, wBuffer, sizeof(wBuffer));
 	return _wchdir(wBuffer);
     }
@@ -2890,7 +2905,7 @@ win32_access(const char *path, int mode)
 {
     dTHXo;
     if (USING_WIDE()) {
-	WCHAR wBuffer[MAX_PATH];
+	WCHAR wBuffer[MAX_PATH+1];
 	A2WHELPER(path, wBuffer, sizeof(wBuffer));
 	return _waccess(PerlDir_mapW(wBuffer), mode);
     }
@@ -2902,7 +2917,7 @@ win32_chmod(const char *path, int mode)
 {
     dTHXo;
     if (USING_WIDE()) {
-	WCHAR wBuffer[MAX_PATH];
+	WCHAR wBuffer[MAX_PATH+1];
 	A2WHELPER(path, wBuffer, sizeof(wBuffer));
 	return _wchmod(PerlDir_mapW(wBuffer), mode);
     }
@@ -3405,7 +3420,7 @@ win32_dynaload(const char* filename)
     dTHXo;
     HMODULE hModule;
     if (USING_WIDE()) {
-	WCHAR wfilename[MAX_PATH];
+	WCHAR wfilename[MAX_PATH+1];
 	A2WHELPER(filename, wfilename, sizeof(wfilename));
 	hModule = LoadLibraryExW(PerlDir_mapW(wfilename), NULL, LOAD_WITH_ALTERED_SEARCH_PATH);
     }
@@ -3839,15 +3854,15 @@ XS(w32_CopyFile)
     if (items != 3)
 	Perl_croak(aTHX_ "usage: Win32::CopyFile($from, $to, $overwrite)");
     if (USING_WIDE()) {
-	WCHAR wSourceFile[MAX_PATH];
-	WCHAR wDestFile[MAX_PATH];
+	WCHAR wSourceFile[MAX_PATH+1];
+	WCHAR wDestFile[MAX_PATH+1];
 	A2WHELPER(SvPV_nolen(ST(0)), wSourceFile, sizeof(wSourceFile));
 	wcscpy(wSourceFile, PerlDir_mapW(wSourceFile));
 	A2WHELPER(SvPV_nolen(ST(1)), wDestFile, sizeof(wDestFile));
 	bResult = CopyFileW(wSourceFile, PerlDir_mapW(wDestFile), !SvTRUE(ST(2)));
     }
     else {
-	char szSourceFile[MAX_PATH];
+	char szSourceFile[MAX_PATH+1];
 	strcpy(szSourceFile, PerlDir_mapA(SvPV_nolen(ST(0))));
 	bResult = CopyFileA(szSourceFile, PerlDir_mapA(SvPV_nolen(ST(1))), !SvTRUE(ST(2)));
     }
